Restructure Core and MVVM folders
