+++ Core/Utilities/ModNameHelper.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Modding_Assistant.Core
+namespace Modding_Assistant.Core.Utilities
 {
     public static class ModNameHelper
     {
