+++ MVVM/Base/RelayCommand.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Input;
 
-namespace Modding_Assistant.Core
+namespace Modding_Assistant.MVVM.Base
 {
     public class RelayCommand(Action<object?> execute, Func<object?, bool>? canExecute = null) : ICommand
     {
@@ -15,11 +15,11 @@ namespace Modding_Assistant.Core
 
         public bool CanExecute(object? parameter)
         {
-            return this.canExecute == null || this.canExecute(parameter);
+            return canExecute == null || canExecute(parameter);
         }
         public void Execute(object? parameter)
         {
-            this.execute(parameter);
+            execute(parameter);
         }
         public void RaiseCanExecuteChanged()
         {
