+++ MVVM/ViewModel/MainViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using ClosedXML.Excel;
 using Microsoft.EntityFrameworkCore;
-using Modding_Assistant.Core;
+using Modding_Assistant.Core.Utilities;
+using Modding_Assistant.MVVM.Base;
 using Modding_Assistant.MVVM.Model;
 using Modding_Assistant.MVVM.Services.Implementations;
 using Modding_Assistant.MVVM.Services.Interfaces;
